Added money to player when he wins game, and upfated money to use int

+++ Assets/Scripts/Gadgets/Gadgets/Gadget.cs
@@ -62,7 +62,7 @@ public abstract class Gadget
         return this.gadgetInfo.name;
     }
 
-    public float getCost()
+    public int getCost()
     {
         return this.gadgetInfo.cost;
     }
